refactor: update and refactor (#36)
* tmp env update

* chore: rename charts

* chore: rename SIdebar

* chore: rename Statistics

* chore: rename Users

* chore: move files

* docs: update readme
