+++ .env
@@ -1,2 +1,2 @@
-VITE_STATS_API_ENDPOINT=http://localhost:3000
+VITE_STATS_API_ENDPOINT=https://vercel-stats-ochre.vercel.app
 VITE_USERS_API_ENDPOINT=https://dummyjson.com/
