+++ src/components/SnowfallToggle.tsx
@@ -2,7 +2,7 @@ import { Button } from '@/components/ui/button';
 import { SnowflakeIcon } from 'lucide-react';
 import SimplelTooltip from './SimplelTooltip';
 import { cn } from '@/lib/utils';
-import { useSnowfall } from '@/features/Sidebar/SidebarProvider';
+import { useSnowfall } from '@/features/sidebar/SidebarProvider';
 
 export default function SnowfallToggle() {
   const { toggleSnowfall, isSnowfallEnabled } = useSnowfall();
