+++ src/features/charts/components/LineChart.tsx
@@ -5,10 +5,10 @@ import {
   ChartTooltip,
   ChartTooltipContent,
 } from '@/components/ui/chart';
-import useChartData from '../../hooks/useChartData';
-import { chartConfig } from '../../config';
+import useChartData from '../hooks/useChartData';
+import { chartConfig } from '../config';
 import { cn } from '@/lib/utils';
-import type { Data } from '../../types';
+import type { Data } from '../types';
 
 export type ChartProps = {
   className?: string;
