+++ src/features/charts/config.ts
