+++ src/pages/Home.tsx
@@ -1,15 +1,15 @@
 import { Button } from '@/components/ui/button';
 import { Skeleton } from '@/components/ui/skeleton';
-import CardLoader from '@/features/Statistics/components/CardLoader';
-import MainStatistics from '@/features/Statistics/MainStatistics';
+import CardLoader from '@/features/statistics/components/CardLoader';
+import MainStatistics from '@/features/statistics/MainStatistics';
 import { lazy, Suspense } from 'react';
 import { Link } from 'react-router';
 
 const DynamicChart = lazy(
-  () => import('@/features/Charts/components/DynamicChart')
+  () => import('@/features/charts/components/DynamicChart')
 );
 const RecentUsers = lazy(
-  () => import('@/features/Users/components/RecentUsers')
+  () => import('@/features/users/components/RecentUsers')
 );
 
 export default function Home() {
