+++ src/pages/charts/ChartsPage.tsx
@@ -1,11 +1,11 @@
-import ChartCard from '@/features/Charts/ChartCard';
+import ChartCard from '@/features/charts/ChartCard';
 import { lazy } from 'react';
 
 const LineChartComponent = lazy(
-  () => import('@/features/Charts/components/LineChart/LineChart')
+  () => import('@/features/charts/components/LineChart')
 );
 const BarChartComponent = lazy(
-  () => import('@/features/Charts/components/BarChart/BarChart')
+  () => import('@/features/charts/components/BarChart')
 );
 
 export default function ChartsPage() {
